refactor: shared status/confidence vocabulary for the screens stack
New leaf module src/lib/screenStatusShared.ts consolidates the duplicated
enums and constant maps across the handoff/trace/export layers:

- TraceConfidence (5-state) moved from screenArtifactTraceBridge (re-exported
  there so importers don't churn); the two byte-identical rank maps and two
  label maps collapse to one TRACE_CONFIDENCE_RANK / TRACE_CONFIDENCE_LABELS.
- ScreenHandoffStatus and ScreensExportStatus defined once;
  ScreenImplementationReadiness / ScreensHandoffExportStatus stay as aliases
  so call-site vocabulary is unchanged. toExportStatus() replaces the inline
  blocked->not_ready mapping in deriveScreensExportStatus.
- TraceabilityConfidence in screenReadiness is now Extract<TraceConfidence,...>
  (type-level anchoring only — the 3-state and 5-state systems stay
  behaviorally separate).
- SystemReadinessStatus deliberately stays in screenReviewWorkflow (persisted
  reviewStatus-adjacent vocabulary) with a pointer comment.

No string value changed anywhere; exported markdown/tests are byte-identical.

7 files changed, +81 / -36.
